Adding "to" operation to variables.
This feature would allow variables to have reassigned values, rather than using the = sign to completely declare the name with a value again. This <<might>> result into faster execution times. The to feature is mainly significant for functions as it allows you to reassign the values of items passed into the function.